change some byteswapping

+++ client_libs/c++/playerclient.h
@@ -836,7 +836,7 @@ class PositionProxy : public ClientProxy
         respectively.\\
         Returns: 0 if everything's ok, -1 otherwise.
     */
-    int SetSpeed(short speed, short turnrate);
+    int SetSpeed(int speed, int turnrate);
 
     /** Enable/disable the motors.
         Set {\tt state} to 0 to disable (default) or 1 to enable.
